+++ src/types/state_variable.cpp
@@ -198,7 +198,7 @@ namespace ratio
 
         for (const auto &[sv, atms] : sv_instances)
         {
-            json::json tl{{"id", get_id(*sv)}, {"type", STATE_VARIABLE_TYPE_NAME}};
+            json::json tl{{"id", static_cast<uint64_t>(get_id(*sv))}, {"type", STATE_VARIABLE_TYPE_NAME}};
 #ifdef COMPUTE_NAMES
             tl["name"] = get_solver().guess_name(*sv);
 #endif
